start moving non-interface things out of pq_helper
