+++ CMakeLists.txt
@@ -43,4 +43,5 @@ set(PYBIND11_CPP_STANDARD -std=c++14)
 pybind11_add_module(_pdaggerq
                     pdaggerq/tensor.cc
                     pdaggerq/pq.cc 
+                    pdaggerq/pq_utils.cc
                     pdaggerq/pq_helper.cc)
