+++ pdaggerq/pq_utils.cc
@@ -1,0 +1,62 @@
+//
+// pdaggerq - A code for bringing strings of creation / annihilation operators to normal order.
+// Filename: pq_helper.cc
+// Copyright (C) 2020 A. Eugene DePrince III
+//
+// Author: A. Eugene DePrince III <adeprince@fsu.edu>
+// Maintainer: DePrince group
+//
+// This file is part of the pdaggerq package.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+#include "pq_utils.h"
+
+namespace pdaggerq {
+
+/// concatinate a list of operators (a list of strings) into a single list
+std::vector<std::string> concatinate_operators(std::vector<std::vector<std::string>> ops) {
+
+    std::vector<std::string> ret;
+    size_t size = 0;
+    for (size_t i = 0; i < ops.size(); i++) {
+        size += ops[i].size();
+    }
+    ret.reserve(size);
+    for (size_t i = 0; i < ops.size(); i++) {
+        ret.insert(ret.end(), ops[i].begin(), ops[i].end());
+    }
+    return ret;
+
+}
+
+/// remove "*" from std::string
+void removeStar(std::string &x) {
+
+  auto it = std::remove_if(std::begin(x),std::end(x),[](char c){return (c == '*');});
+  x.erase(it, std::end(x));
+}
+
+/// remove "(" and ")" from std::string
+void removeParentheses(std::string &x) {
+
+  auto it = std::remove_if(std::begin(x),std::end(x),[](char c){return (c == '(');});
+  x.erase(it, std::end(x));
+
+  it = std::remove_if(std::begin(x),std::end(x),[](char c){return (c == ')');});
+  x.erase(it, std::end(x));
+
+}
+
+}
