+++ pdaggerq/pq_utils.h
@@ -1,0 +1,43 @@
+//
+// pdaggerq - A code for bringing strings of creation / annihilation operators to normal order.
+// Filename: pq_helper.cc
+// Copyright (C) 2020 A. Eugene DePrince III
+//
+// Author: A. Eugene DePrince III <adeprince@fsu.edu>
+// Maintainer: DePrince group
+//
+// This file is part of the pdaggerq package.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+#ifndef PQ_UTILS_H
+#define PQ_UTILS_H
+
+#include<string>
+#include<vector>
+
+namespace pdaggerq {
+
+/// concatinate a list of operators (a list of strings) into a single list
+std::vector<std::string> concatinate_operators(std::vector<std::vector<std::string>> ops);
+
+/// remove "*" from std::string
+void removeStar(std::string &x);
+
+/// remove "(" and ")" from std::string
+void removeParentheses(std::string &x);
+
+}
+
+#endif
